Added segmentSize for topic level

+++ kafka-ui-api/src/main/java/com/provectus/kafka/ui/cluster/model/InternalTopic.java
@@ -20,6 +20,6 @@ public class InternalTopic {
     private final int replicationFactor;
     private final int underReplicatedPartitions;
     //TODO: find way to fill
-    private final int segmentSize;
+    private final long segmentSize;
     private final int segmentCount;
 }
